# openapi-schema.yaml
openapi: 3.1.0
info:
  title: Správa Bytov API
  description: REST API pre správu bytov, nájomníkov a finančných štatistík cez Google Sheets
  version: 1.0.0

servers:
  - url: https://sprava-bytov-production.up.railway.app
    description: Production server (Railway)

paths:
  /apartments:
    get:
      summary: Získať všetky byty
      description: Vráti zoznam všetkých bytov z Google Sheets
      operationId: getApartments
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Zoznam bytov
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Apartment'
    post:
      summary: Pridať nový byt
      description: Pridá nový byt do Google Sheets
      operationId: addApartment
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApartmentInput'
      responses:
        '201':
          description: Byt úspešne pridaný

  /apartments/{id}:
    get:
      summary: Získať konkrétny byt
      description: Vráti detail bytu podľa Byt_ID
      operationId: getApartmentById
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID bytu (napr. BYT-001)
      responses:
        '200':
          description: Detail bytu
        '404':
          description: Byt nenájdený
    patch:
      summary: Aktualizovať byt
      description: Aktualizuje vybrané polia konkrétneho bytu
      operationId: updateApartment
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApartmentUpdate'
      responses:
        '200':
          description: Byt úspešne aktualizovaný

  /stats/summary:
    get:
      summary: Získať finančné štatistiky
      description: Vráti agregované štatistiky všetkých bytov (cashflow, výnosnosť, náklady)
      operationId: getStatsSummary
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Finančné štatistiky
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsSummary'

  /tenants:
    get:
      summary: Získať všetkých nájomníkov
      description: Vráti zoznam všetkých nájomníkov s informáciami o byte
      operationId: getTenants
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Zoznam nájomníkov
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'

  /tenants/{name}:
    get:
      summary: Vyhľadať nájomníka
      description: Vyhľadá nájomníka podľa mena (case-insensitive)
      operationId: searchTenant
      security:
        - ApiKeyAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Meno alebo časť mena nájomníka
      responses:
        '200':
          description: Nájdený nájomník/nájomníci
        '404':
          description: Nájomník nenájdený

  /columns:
    get:
      summary: Získať zoznam všetkých stĺpcov v tabuľke
      operationId: getColumns
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Zoznam stĺpcov
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: Počet stĺpcov
                  columns:
                    type: array
                    items:
                      type: string
                    description: Názvy stĺpcov
        '500':
          description: Chyba servera
    
    post:
      summary: Pridať nový stĺpec do tabuľky
      operationId: addColumn
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - columnName
              properties:
                columnName:
                  type: string
                  description: Názov nového stĺpca
              example:
                columnName: "Čistá výnosnosť"
      responses:
        '201':
          description: Stĺpec bol úspešne pridaný
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  columnName:
                    type: string
        '400':
          description: Neplatný request
        '409':
          description: Stĺpec už existuje
        '500':
          description: Chyba servera

  /columns/{oldName}:
    patch:
      summary: Premenovať existujúci stĺpec
      operationId: renameColumn
      security:
        - ApiKeyAuth: []
      parameters:
        - name: oldName
          in: path
          required: true
          schema:
            type: string
          description: Aktuálny názov stĺpca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newName
              properties:
                newName:
                  type: string
                  description: Nový názov stĺpca
              example:
                newName: "Ročná výnosnosť (%)"
      responses:
        '200':
          description: Stĺpec bol úspešne premenovaný
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  oldName:
                    type: string
                  newName:
                    type: string
        '400':
          description: Neplatný request
        '404':
          description: Stĺpec nebol nájdený
        '500':
          description: Chyba servera

  /columns/{columnName}:
    delete:
      summary: Vymazať obsah stĺpca
      operationId: deleteColumn
      security:
        - ApiKeyAuth: []
      parameters:
        - name: columnName
          in: path
          required: true
          schema:
            type: string
          description: Názov stĺpca na vymazanie
      responses:
        '200':
          description: Stĺpec bol úspešne vymazaný
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  columnName:
                    type: string
        '404':
          description: Stĺpec nebol nájdený
        '500':
          description: Chyba servera

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key

  schemas:
    Apartment:
      type: object
      properties:
        Byt_ID:
          type: string
          description: Unikátne ID bytu (napr. BYT-001)
        Adresa_ulica:
          type: string
        Mesto:
          type: string
        PSC:
          type: string
        Typ_bytu:
          type: string
        Rozloha_m2:
          type: number
          nullable: true
        Kupna_cena_EUR:
          type: number
          nullable: true
        Najom_brutto_EUR:
          type: number
          nullable: true
        Najomnik_meno:
          type: string
        Najomnik_telefon:
          type: string
        Najomnik_email:
          type: string

    ApartmentInput:
      type: object
      properties:
        Adresa_ulica:
          type: string
          description: Ulica a číslo
        Mesto:
          type: string
          description: Mesto
        PSC:
          type: string
          description: PSČ
        Typ_bytu:
          type: string
          description: Typ bytu (napr. 2-izbový)
        Rozloha_m2:
          type: number
          description: Rozloha v m2
        Kupna_cena_EUR:
          type: number
          description: Kúpna cena v EUR
        Vlastne_zdroje_EUR:
          type: number
          description: Vlastné zdroje v EUR
        Uver_vyska_EUR:
          type: number
          description: Výška úveru v EUR
        Urokova_sadzba:
          type: number
          description: Úroková sadzba v %
        Pocet_rokov_uveru:
          type: number
          description: Počet rokov úveru
        Banka:
          type: string
          description: Názov banky
        Mesacna_splatka_EUR:
          type: number
          description: Mesačná splátka v EUR
        Druzstvo_EUR:
          type: number
          description: Mesačné náklady na družstvo v EUR
        Elektrina_EUR:
          type: number
          description: Mesačné náklady na elektrinu v EUR
        Plyn_EUR:
          type: number
          description: Mesačné náklady na plyn v EUR
        Najom_brutto_EUR:
          type: number
          description: Mesačný nájom brutto v EUR
        Najomnik_meno:
          type: string
          description: Meno nájomníka

    ApartmentUpdate:
      type: object
      properties:
        Adresa_ulica:
          type: string
        Mesto:
          type: string
        PSC:
          type: string
        Typ_bytu:
          type: string
        Rozloha_m2:
          type: number
        Kupna_cena_EUR:
          type: number
        Vlastne_zdroje_EUR:
          type: number
        Uver_vyska_EUR:
          type: number
        Urokova_sadzba_p_a_percent:
          type: number
        Pocet_rokov_uveru:
          type: number
        Banka:
          type: string
        Datum_zaciatku_uveru:
          type: string
        Mesacna_splatka_uveru_EUR:
          type: number
        Druzstvo_EUR:
          type: number
        Elektrina_EUR:
          type: number
        Plyn_EUR:
          type: number
        Voda_EUR:
          type: number
        Internet_EUR:
          type: number
        Ostatne_naklady_EUR:
          type: number
        Celkove_mesacne_naklady_bez_uveru_EUR:
          type: number
        Najom_brutto_EUR:
          type: number
        Cashflow_mesacne_EUR:
          type: number
        Rocna_hruba_vynosnost_percent:
          type: number
        Najomnik_meno:
          type: string
        Najomnik_telefon:
          type: string
        Najomnik_email:
          type: string
        Zmluva_od:
          type: string
        Zmluva_do:
          type: string
        Vyska_depozitu_EUR:
          type: number
        Stav_platieb:
          type: string
        Poznamka:
          type: string

    StatsSummary:
      type: object
      properties:
        count:
          type: integer
          description: Počet bytov
        totalPurchase:
          type: number
          description: Súčet kúpnych cien v EUR
        totalMonthlyCosts:
          type: number
          description: Súčet mesačných nákladov v EUR
        totalMonthlyRent:
          type: number
          description: Súčet mesačných nájmov v EUR
        totalMonthlyCashflow:
          type: number
          description: Súčet mesačného cashflow v EUR
        averageYield:
          type: number
          nullable: true
          description: Priemerná ročná výnosnosť v %

    Tenant:
      type: object
      properties:
        meno:
          type: string
        telefon:
          type: string
        email:
          type: string
        byt_id:
          type: string
        vyska_najomu:
          type: number
        mesacne_naklady_bytu:
          type: number